requirements.txt
streamlit>=1.35.0
chromadb>=0.5.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
cryptography>=42.0.0
werkzeug>=3.0.0
openai>=1.0.0
anthropic>=0.25.0
python-docx>=1.1.0
tenacity>=8.0.0
